total refactor of client code
