+++ include/chatp2p/error.h
@@ -19,17 +19,17 @@
 #define CHAT_ERROR_H
 
 typedef enum {
-	CHAT_BAD_ADDRESS,
-	CHAT_NO_ENTRY,
-	CHAT_DECRYPT_ERROR,
-	CHAT_INVALID_CONNECT_MSG,
-	CHAT_ALREADY_CONNECTED,
-	CHAT_ADDR_BOOK_ERROR,
-	CHAT_CONNECTION_DENIED,
-	CHAT_BAD_PING,
-	CHAT_INVALID_SERVER_KEY,
-	CHAT_INVALID_MESSAGE_TYPE,
-	CHAT_ERROR_COUNT
+    CHAT_BAD_ADDRESS,
+    CHAT_NO_ENTRY,
+    CHAT_DECRYPT_ERROR,
+    CHAT_INVALID_CONNECT_MSG,
+    CHAT_ALREADY_CONNECTED,
+    CHAT_ADDR_BOOK_ERROR,
+    CHAT_CONNECTION_DENIED,
+    CHAT_BAD_PING,
+    CHAT_INVALID_SERVER_KEY,
+    CHAT_INVALID_MESSAGE_TYPE,
+    CHAT_ERROR_COUNT
 } ChatErrorType;
 
 typedef void MessageErrorHandler(void *data);
